MANATEE-203 fix bug in handling of shared buffers in config

+++ boot/configure.sh
@@ -10,7 +10,11 @@ set -o xtrace
 # set shared_buffers to 1/4 provisoned RSS
 set -o errexit
 set -o pipefail
-cp /opt/smartdc/manatee/etc/postgresql.sdc.conf /opt/smartdc/manatee/etc/postgresql.sdc.conf.in
+# make a backup if non exists.
+if [[ ! -f /opt/smartdc/manatee/etc/postgresql.sdc.conf.in ]]
+then
+    cp /opt/smartdc/manatee/etc/postgresql.sdc.conf /opt/smartdc/manatee/etc/postgresql.sdc.conf.in
+fi
 shared_buffers="$(( $(prtconf -m) / 4 ))MB"
 sed -e "s#@@SHARED_BUFFERS@@#$shared_buffers#g" \
     /opt/smartdc/manatee/etc/postgresql.sdc.conf.in > /opt/smartdc/manatee/etc/postgresql.sdc.conf
